refactor: move eventsPlugin
