+++ plugins/events/package.json
@@ -58,6 +58,7 @@
 		"eventemitter3": "^5.0.0"
 	},
 	"devDependencies": {
+		"@just-web/app": "workspace:^",
 		"@just-web/log": "workspace:^",
 		"@just-web/repo-scripts": "workspace:^",
 		"@just-web/testing": "workspace:^",
@@ -82,6 +83,7 @@
 		"typescript": "^5.0.0"
 	},
 	"peerDependencies": {
+		"@just-web/app": "workspace:^",
 		"@just-web/log": "workspace:^"
 	},
 	"typedoc": {
